// src/api/models/TenantUpdateReq.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TenantUpdateReq = {
    /**
     * 企业地址
     */
    company_address?: string;
    /**
     * 公司名称
     */
    company_name?: string;
    /**
     * 企业logo
     */
    logo_address?: string;
    /**
     * 租户id
     */
    tenant_id?: number;
};